add:getmapping

+++ src/main/kotlin/com/example/todoApp/dto/TodoResponse.kt
@@ -1,3 +1,3 @@
 package com.example.todoApp.dto
 
-data class TodoResponse(val id: Int, val text: String)
+data class TodoResponse(val id: Long, val text: String)
